+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * improved default title
 * only execute handlers when the theme is active
 * modified the handler so that redirect which correct the URI still retain the query string
+* do not set a canonical URI on error
 
 ## 0.1a12 (19.10.2018)
 ### Features